feat(customers): move balance billing date into edit sheet
Move the billing-cycle "Bills" date out of the balance table into the
edit sheet as an InfoRow. In the table, keep the Resets chip and show an
amber clock icon (with expiry tooltip) in a fixed left slot so usage bars
stay aligned across rows.
